requirements.txt
flask
python-dotenv
psycopg2-binary
flask-swagger
flask-swagger-ui
flask-sqlalchemy